pyproject.toml
[project]
name = "mongodb"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabriel Martin del Campo y Eternod", email = "gmartindelcye@gmail.com"},
]
dependencies = [
    "pymongo>=4.6.2",
    "pymongoose>=1.3.7",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
